Add tpt-gpu-doctor, hosted playground build, and dependency-hygiene CI
Adds a doctor CLI mirroring CI's fmt/clippy/toolchain checks with
--pre-commit/--fast flags, wires cargo-deny/cargo-audit + Dependabot
into CI, deploys the wasm playground via docs.yml instead of requiring
a local wasm-pack build, and marks JAX integration as an honest stub
rather than complete in todo.md/CLAUDE.md. Remaining changes are
formatting/cleanup across existing modules.

+++ crates/tpt-gpu-runtime/src/device/device.rs
@@ -575,7 +575,7 @@ impl Device {
                             "reduce_max: input buffer unavailable",
                         )
                     })?;
-                    let elems = (guard.len() / 4) as usize;
+                    let elems = guard.len() / 4;
                     let mut m = f32::NEG_INFINITY;
                     for i in 0..elems {
                         let off = i * 4;
@@ -618,6 +618,7 @@ impl Device {
         Ok(())
     }
     /// Lock the backing arena buffer for a live device pointer.
+    #[allow(dead_code)]
     fn lock_ptr(&self, device_ptr: u64) -> TptrResult<(u64, MutexGuard<'_, Vec<u8>>)> {
         let guard = self.arena.get(device_ptr)?;
         let size = guard.len() as u64;
